Unit test fixes
